requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
black==21.12b0
click==8.0.3
Django==4.0.1
djangorestframework==3.13.1
flake8==4.0.1
importlib-metadata==4.10.0
Markdown==3.3.6
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyright==0.0.13
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
zipp==3.7.0
